refactor: split collection module into focused submodules
Split the collection module into logical submodules for better
maintainability and clarity:

- collection/error.rs: Error types and collection outcomes
- collection/types.rs: Core traits and types (Collector trait, Location)
- collection/config.rs: Collection configuration
- collection/nodes.rs: Node implementations (Session, Directory, Module, Function)
- collection/utils.rs: Utility functions (glob matching)

This refactoring improves code organization and makes the codebase
easier to navigate and maintain.
