+++ rtest-core/src/collection/mod.rs
@@ -1,0 +1,9 @@
+//! Pytest collection implementation.
+//!
+//! This module provides a modular implementation of pytest's collection logic.
+
+pub mod config;
+pub mod error;
+pub mod nodes;
+pub mod types;
+mod utils;
